Add Available to ResourceInfo struct
Add a new field named 'Available' to the ResourceInfo structure in order
to align with the terminology used in node status. Now ResourceInfo
is able to carry the same information as node status plus the amount of
resource currently available for new pods to be allocated (which is not
present in node status).

Also add comments describing these fields for future readers.